chore(compose): split the ops seams under the file-length cap
