+++ backend/internal/compose/attentionseam.go
@@ -376,22 +376,6 @@ func (f attentionFailedEffects) Failed(ctx context.Context, limit int) ([]attent
 	return out, nil
 }
 
-// attentionDSRs binds the compliance lane to the consent module's own thin
-// read; the DSR-admin gate lives in the store, not here.
-type attentionDSRs struct{ store *consent.Store }
-
-func (d attentionDSRs) OpenDueSoonest(ctx context.Context, limit int) ([]attention.DSRCase, error) {
-	owed, err := d.store.OpenDSRsDueSoonest(ctx, limit)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]attention.DSRCase, 0, len(owed))
-	for _, request := range owed {
-		out = append(out, attention.DSRCase{ID: request.ID, Kind: request.Kind, DueAt: request.DueAt})
-	}
-	return out, nil
-}
-
 // attentionBriefing binds the briefing lane to the same engine entry point Home
 // and the agent tool read, so all three read one queue rather than three
 // readings of it.
@@ -433,29 +417,6 @@ func (a attentionBriefing) Queue(ctx context.Context) ([]attention.BriefEntry, b
 		})
 	}
 	return entries, true, nil
-}
-
-// attentionSyncHealth binds the sync-health lane to the overlay module's own
-// aggregated read; the mode gate and the every-role read posture live there.
-type attentionSyncHealth struct{ svc *overlay.Service }
-
-func (h attentionSyncHealth) Concerns(ctx context.Context) ([]attention.SyncConcern, error) {
-	concerns, err := h.svc.SyncHealth(ctx)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]attention.SyncConcern, 0, len(concerns))
-	for _, concern := range concerns {
-		out = append(out, attention.SyncConcern{
-			Kind:        concern.Kind,
-			ErrorClass:  concern.ErrorClass,
-			Failures:    concern.Failures,
-			NextSweepAt: concern.NextSweepAt,
-			Band:        concern.Band,
-			Objects:     concern.Objects,
-		})
-	}
-	return out, nil
 }
 
 // newAttentionHandlers assembles the surface for the API role. meter is the
